+++ docs/_templates/components/grid.blade.php
@@ -1,0 +1,3 @@
+<div class="grid grid-cols-3 gap-4">
+    {{$slot}}
+</div>