+++ examples/basic-page-builder/resources/views/site/blocks/text.blade.php
@@ -1,0 +1,4 @@
+<div class="prose">
+    <h2>{{$block->translatedInput('title')}}</h2>
+    {!! $block->translatedInput('text') !!}
+</div>