+++ src/Exceptions/Handler.php
@@ -18,7 +18,7 @@ class Handler extends ExceptionHandler
     protected function getHttpExceptionView(HttpExceptionInterface $e)
     {
         $usesAdminPath = !empty(config('twill.admin_app_path'));
-        $adminAppUrl = config('twill.admin_app_url');
+        $adminAppUrl = config('twill.admin_app_url', config('app.url'));
 
         $isSubdomainAdmin = !$usesAdminPath && Str::contains(Request::url(), $adminAppUrl);
         $isSubdirectoryAdmin = $usesAdminPath && Str::startsWith(Request::path(), config('twill.admin_app_path'));