+++ src/Http/Controllers/Admin/ModuleController.php
@@ -13,6 +13,7 @@ use A17\Twill\Models\Contracts\TwillSchedulableModel;
 use A17\Twill\Models\Group;
 use A17\Twill\Repositories\ModuleRepository;
 use A17\Twill\Services\Blocks\Block;
+use A17\Twill\Services\Breadcrumbs\Breadcrumbs;
 use A17\Twill\Services\Forms\Form;
 use A17\Twill\Services\Listings\Columns\Browser;
 use A17\Twill\Services\Listings\Columns\FeaturedStatus;
@@ -343,6 +344,11 @@ abstract class ModuleController extends Controller
     protected $labels = [];
 
     /**
+     * When set to true and the model is translatable, the language prefix will not be shown in the permalink.
+     */
+    private bool $withoutLanguageInPermalink = false;
+
+    /**
      * The columns to search for when using the search field.
      *
      * Do not modify this directly but use the method setSearchColumns().
@@ -364,6 +370,8 @@ abstract class ModuleController extends Controller
             ],
         ],
     ];
+
+    private ?Breadcrumbs $breadcrumbs = null;
 
     public function __construct(Application $app, Request $request)
     {
@@ -613,6 +621,11 @@ abstract class ModuleController extends Controller
         $this->permalinkBase = $permalinkBase;
     }
 
+    protected function withoutLanguageInPermalink(bool $without = true): void
+    {
+        $this->withoutLanguageInPermalink = $without;
+    }
+
     /**
      * Sets the field to use as title, defaults to `title`.
      */
@@ -665,6 +678,14 @@ abstract class ModuleController extends Controller
     }
 
     /**
+     * Set the breadcrumbs.
+     */
+    protected function setBreadcrumbs(Breadcrumbs $breadcrumbs): void
+    {
+        $this->breadcrumbs = $breadcrumbs;
+    }
+
+    /**
      * $type can be index or browser.
      */
     private function getTableColumns(string $type): TableColumns
@@ -676,7 +697,18 @@ abstract class ModuleController extends Controller
         }
 
         return $tableColumns->each(function (TableColumn $column) {
-            if ($column->shouldLinkToEdit()) {
+            if ($column instanceof NestedData) {
+                $column->linkCell(function (TwillModelContract $model, NestedData $column) {
+                    $module = Str::singular(last(explode('.', $this->moduleName)));
+
+                    return moduleRoute(
+                        "$this->moduleName." . $column->getField(),
+                        $this->routePrefix,
+                        'index',
+                        [$module => $this->getItemIdentifier($model)]
+                    );
+                });
+            } elseif ($column->shouldLinkToEdit()) {
                 $column->linkCell(function (TwillModelContract $model) {
                     if ($model->trashed()) {
                         return null;
@@ -1105,7 +1137,7 @@ abstract class ModuleController extends Controller
 
         return View::make($view, $this->form($id))->with(
             'renderFields',
-            $this->getForm($this->repository->getById($id))
+            $controllerForm
         );
     }
 
@@ -1691,6 +1723,13 @@ abstract class ModuleController extends Controller
                 unset($indexDataWithoutFilters[$key]);
             }
         }
+
+        if ($this->breadcrumbs && !isset($indexDataWithoutFilters['breadcrumb'])) {
+            foreach ($this->breadcrumbs->getListingBreadcrumbs() as $breadcrumb) {
+                $indexDataWithoutFilters['breadcrumb'][] = $breadcrumb->toArray();
+            }
+        }
+
         $filters = $this->filters()->toFrontendArray($this->repository);
 
         return array_replace_recursive($data + $options, $indexDataWithoutFilters + $filters);
@@ -2190,6 +2229,12 @@ abstract class ModuleController extends Controller
 
         $form = array_replace_recursive($data, $this->formData($this->request));
 
+        if ($this->breadcrumbs && !isset($form['breadcrumb'])) {
+            foreach ($this->breadcrumbs->getFormBreadcrumbs() as $breadcrumb) {
+                $form['breadcrumb'][] = $breadcrumb->toArray();
+            }
+        }
+
         View::share('form', $form);
 
         return $form;
@@ -2393,7 +2438,7 @@ abstract class ModuleController extends Controller
         }
 
         return $appUrl . '/'
-            . ($this->moduleHas('translations') ? '{language}/' : '')
+            . ((!$this->withoutLanguageInPermalink && $this->moduleHas('translations')) ? '{language}/' : '')
             . ($this->moduleHas('revisions') ? '{preview}/' : '')
             . (empty($this->getLocalizedPermalinkBase()) ? ($this->permalinkBase ?? $this->getModulePermalinkBase(
             )) : '')
@@ -2405,7 +2450,7 @@ abstract class ModuleController extends Controller
     /**
      * @return array
      */
-    protected function getLocalizedPermalinkBase()
+    protected function getLocalizedPermalinkBase(): array
     {
         return [];
     }