+++ src/Services/Forms/Options.php
@@ -5,7 +5,7 @@ namespace A17\Twill\Services\Forms;
 use Illuminate\Support\Collection;
 
 /**
- * @implements Collection<int, Option>
+ * @implements Collection<int, Options>
  */
 class Options extends Collection
 {